Separate Modal component from modal contents
The original design had only one modal so the modal design was wrapped
up with the contents shown in the modal. Here we separate them to allow
different modals in future commits.

There should be no user-facing changes with this change.
