+++ src/actions/types.js
@@ -28,8 +28,7 @@ export const ADD_NOTIFICATION = "ADD_NOTIFICATION";
 export const REMOVE_NOTIFICATION = "REMOVE_NOTIFICATION";
 export const CHANGE_PANEL_LAYOUT = "CHANGE_PANEL_LAYOUT";
 export const TOGGLE_PANEL_DISPLAY = "TOGGLE_PANEL_DISPLAY";
-export const TRIGGER_DOWNLOAD_MODAL = "TRIGGER_DOWNLOAD_MODAL";
-export const DISMISS_DOWNLOAD_MODAL = "DISMISS_DOWNLOAD_MODAL";
+export const SET_MODAL = "SET_MODAL";
 export const ADD_EXTRA_METADATA = "ADD_EXTRA_METADATA";
 export const CHANGE_TREE_ROOT_IDX = "CHANGE_TREE_ROOT_IDX";
 export const TOGGLE_NARRATIVE = "TOGGLE_NARRATIVE";
